fix: onboarding 수정
- 학생, 강사 컴포넌트 임시 분리
- domain 값 수정
- useOnboardingStatus를 학생 전용으로 전환
