+++ src/features/dashboard/hooks/use-onboarding-query.ts
@@ -3,10 +3,11 @@ import { useQuery } from '@tanstack/react-query';
 
 const ONBOARDING_CHECK_INTERVAL = 3 * 1000; // 3초
 
-export const useTeacherOnboardingQuery = () =>
+export const useTeacherOnboardingQuery = (options?: { enabled?: boolean }) =>
   useQuery({
     queryKey: onboardingKeys.teacher(),
     queryFn: () => repository.teacher.getTeacherOnboarding(),
     staleTime: ONBOARDING_CHECK_INTERVAL,
     refetchInterval: ONBOARDING_CHECK_INTERVAL,
+    enabled: options?.enabled,
   });
